URL aliasing, also finished chat.
If you're going to be running this site on a non-apache server, the
links where it's aliased aren't going to be functional. I think. Of
course for now this is only in the profile links, but nevertheless

Also i'm practically done with the chat system. Just have to add some
simple UI functionalities and I am golden